refactor(xtask): move ignore categorization into perl-ci-hygiene (#0000)
